resolve imports for what's in place now, untested
